requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asyncpg==0.30.0
attrs==25.3.0
disnake==2.10.1
frozenlist==1.7.0
idna==3.10
multidict==6.5.0
propcache==0.3.2
python-dotenv==1.1.0
yarl==1.20.1
